# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: API segura do Família Segura
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /family:
    get:
      operationId: getFamilyOverview
      tags: [family]
      responses:
        "200":
          description: Family overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyOverview"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createFamily
      tags: [family]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyInput"
      responses:
        "201":
          description: Family created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyOverview"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      operationId: deleteFamily
      tags: [privacy]
      responses:
        "204":
          description: Family deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /family/settings:
    patch:
      operationId: updateFamilySettings
      tags: [family]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FamilySettingsUpdate" }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FamilySettings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /family/guardian-pin:
    put:
      operationId: setGuardianPin
      tags: [family]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GuardianPinInput" }
      responses:
        "204":
          description: PIN atualizado
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /family/children:
    post:
      operationId: createChild
      tags: [family]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChildInput"
      responses:
        "201":
          description: Child created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChildProfile"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/PaymentRequired" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /family/children/{childId}:
    patch:
      operationId: updateChild
      tags: [family]
      parameters:
        - $ref: "#/components/parameters/ChildId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChildUpdate" }
      responses:
        "200":
          description: Child updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChildProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: archiveChild
      tags: [family]
      parameters:
        - $ref: "#/components/parameters/ChildId"
      responses:
        "204":
          description: Perfil arquivado e aparelhos revogados
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /family/invites:
    post:
      operationId: createInvite
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InviteInput" }
      responses:
        "201":
          description: Convite criado (código exibido uma única vez)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invite" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/PaymentRequired" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /family/invites/accept:
    post:
      operationId: acceptInvite
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InviteAcceptInput" }
      responses:
        "200":
          description: Entrou na família
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FamilyOverview" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /family/members/{memberId}:
    patch:
      operationId: updateMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemberUpdate" }
      responses:
        "200":
          description: Papel atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FamilyMember" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: removeMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "204":
          description: Membro removido (ou saiu da família)
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /family/push-tokens:
    post:
      operationId: registerGuardianPushToken
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PushTokenInput" }
      responses:
        "204":
          description: Token registrado
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /family/audit:
    get:
      operationId: listAuditEvents
      tags: [privacy]
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEvent"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /family/export:
    get:
      operationId: exportFamilyData
      tags: [privacy]
      responses:
        "200":
          description: Portable family data export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyExport"
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  parameters:
...
    ChildId:
      in: path
      name: childId
      required: true
      schema: { type: string, format: uuid }
    MemberId:
      in: path
      name: memberId
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PaymentRequired:
      description: Recurso ou limite exclusivo do Premium
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Papel sem permissão para a ação
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Conflito (limite atingido ou registro duplicado)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    TooManyRequests:
      description: Muitas tentativas
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    FamilyInput:
      type: object
      required: [name, guardianName, childName, childBirthYear, consentAccepted]
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        guardianName: { type: string, minLength: 2, maxLength: 80 }
        childName: { type: string, minLength: 1, maxLength: 50 }
        childBirthYear: { type: integer, minimum: 2006, maximum: 2030 }
        consentAccepted: { type: boolean }
        timezone: { type: string, minLength: 3, maxLength: 64 }
    FamilySettings:
      type: object
      required: [timezone, offlineLeaseHours, quarantineNewApps, hasGuardianPin]
      properties:
        timezone: { type: string }
        offlineLeaseHours: { type: integer }
        quarantineNewApps: { type: boolean }
        hasGuardianPin: { type: boolean }
    FamilySettingsUpdate:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        timezone: { type: string, minLength: 3, maxLength: 64 }
        offlineLeaseHours: { type: integer, minimum: 12, maximum: 720 }
        quarantineNewApps: { type: boolean }
...
    GuardianPinInput:
      type: object
      required: [pin]
      properties:
        pin: { type: string, pattern: "^[0-9]{4,8}$" }
    ChildInput:
      type: object
      required: [displayName, birthYear]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 50 }
        birthYear: { type: integer, minimum: 2006, maximum: 2030 }
        color: { type: string, pattern: "^#[0-9A-Fa-f]{6}$" }
    ChildUpdate:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 50 }
        birthYear: { type: integer, minimum: 2006, maximum: 2030 }
        color: { type: string, pattern: "^#[0-9A-Fa-f]{6}$" }
...
    PushTokenInput:
      type: object
      required: [token, platform]
      properties:
        token: { type: string, minLength: 10, maxLength: 300 }
        platform: { type: string, enum: [ios, android] }
...
    InviteInput:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [guardian, viewer] }
    Invite:
      type: object
      required: [code, role, expiresAt]
      properties:
        code: { type: string }
        role: { type: string, enum: [guardian, viewer] }
        expiresAt: { type: string, format: date-time }
    InviteAcceptInput:
      type: object
      required: [code, displayName, consentAccepted]
      properties:
        code: { type: string, minLength: 6, maxLength: 20 }
        displayName: { type: string, minLength: 2, maxLength: 80 }
        consentAccepted: { type: boolean }
    MemberUpdate:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [guardian, viewer] }
...
    Family:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        createdAt: { type: string, format: date-time }
    FamilyMember:
      type: object
      required: [id, displayName, role, isCurrentUser]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        role: { type: string, enum: [owner, guardian, viewer] }
        isCurrentUser: { type: boolean }
    ChildProfile:
      type: object
      required: [id, displayName, birthYear, color, ageBand]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        birthYear: { type: integer }
        color: { type: string }
        ageBand:
          type: string
          enum: [ate9, de10a12, de13a15, de16a17, adulto]
    Device:
      type: object
      required: [id, childId, name, platform, status, protectionState, protectionIssues, lastSeenAt, online]
      properties:
        id: { type: string, format: uuid }
        childId: { type: string, format: uuid }
        name: { type: string }
        platform: { type: string, enum: [ios, android] }
        status: { type: string, enum: [active, offline, revoked] }
        protectionState: { type: string, enum: [unknown, active, partial, disabled, unavailable] }
        protectionIssues:
          type: array
          items: { type: string }
        protectionUpdatedAt: { type: [string, "null"], format: date-time }
        lastSeenAt: { type: string, format: date-time }
        online: { type: boolean }
        osVersion: { type: [string, "null"] }
        model: { type: [string, "null"] }
        appVersion: { type: [string, "null"] }
        batteryLevel: { type: [integer, "null"] }
...
    DeviceApp:
      type: object
      required: [id, deviceId, packageName, label, status, firstSeenAt, lastSeenAt]
      properties:
        id: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        packageName: { type: string }
        label: { type: string }
        status: { type: string, enum: [approved, blocked, pending] }
        installedAt: { type: [string, "null"], format: date-time }
        removedAt: { type: [string, "null"], format: date-time }
        firstSeenAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
...
    DeviceEvent:
      type: object
      required: [id, deviceId, childId, type, occurredAt]
      properties:
        id: { type: string, format: uuid }
        deviceId: { type: string, format: uuid }
        childId: { type: string, format: uuid }
        type: { type: string }
        detail: { type: [string, "null"] }
        occurredAt: { type: string, format: date-time }
    AppRule:
      type: object
      required: [id, childId, appId, appName, category, icon, iconColor, androidPackages, usageTodayMinutes, dailyLimitMinutes, extraTodayMinutes, effectiveLimitMinutes, status]
      properties:
        id: { type: string, format: uuid }
        childId: { type: string, format: uuid }
        appId: { type: string }
        appName: { type: string }
        category: { type: string }
        icon: { type: string }
        iconColor: { type: string }
        androidPackages:
          type: array
          items: { type: string }
        usageTodayMinutes: { type: integer }
        dailyLimitMinutes:
          type: integer
          description: No painel do responsável é o limite base. Na visão da criança já inclui o tempo extra de hoje.
        extraTodayMinutes: { type: integer }
        effectiveLimitMinutes: { type: integer }
        status: { type: string, enum: [allowed, attention, blocked] }
    Routine:
      type: object
      required: [id, childId, title, description, days, startTime, endTime, enabled, icon]
      properties:
        id: { type: string, format: uuid }
        childId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        days: { type: string }
        startTime: { type: string }
        endTime: { type: string }
        enabled: { type: boolean }
        icon: { type: string }
    TimeRequest:
      type: object
      required: [id, childId, childName, appId, appName, requestedMinutes, message, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        childId: { type: string, format: uuid }
        childName: { type: string }
        appId: { type: string }
        appName: { type: string }
        requestedMinutes: { type: integer }
        message: { type: string }
        status: { type: string, enum: [pending, approved, denied] }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: [string, "null"], format: date-time }
    AuditEvent:
      type: object
      required: [id, action, summary, createdAt]
      properties:
        id: { type: string, format: uuid }
        action: { type: string }
        summary: { type: string }
        actorName: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
    PrivacySummary:
      type: object
      required: [consentAcceptedAt, retentionDays, collectedData]
      properties:
        consentAcceptedAt: { type: string, format: date-time }
        retentionDays: { type: integer }
        collectedData:
          type: array
          items: { type: string }
    FamilyOverview:
      type: object
      required: [family, members, children, devices, apps, routines, timeRequests, privacy, limits, settings, pendingApps, recentEvents, today]
      properties:
        family: { $ref: "#/components/schemas/Family" }
        today: { type: string, description: Data de hoje no fuso da família (YYYY-MM-DD) }
        members:
          type: array
          items: { $ref: "#/components/schemas/FamilyMember" }
        children:
          type: array
          items: { $ref: "#/components/schemas/ChildProfile" }
        devices:
          type: array
          items: { $ref: "#/components/schemas/Device" }
        apps:
          type: array
          items: { $ref: "#/components/schemas/AppRule" }
        routines:
          type: array
          items: { $ref: "#/components/schemas/Routine" }
        timeRequests:
          type: array
          items: { $ref: "#/components/schemas/TimeRequest" }
        pendingApps:
          type: array
          items: { $ref: "#/components/schemas/DeviceApp" }
        recentEvents:
          type: array
          items: { $ref: "#/components/schemas/DeviceEvent" }
        limits: { $ref: "#/components/schemas/FamilyLimits" }
        settings: { $ref: "#/components/schemas/FamilySettings" }
        privacy: { $ref: "#/components/schemas/PrivacySummary" }
...
    FamilyExport:
      type: object
      required: [exportedAt, data, auditEvents]
      properties:
        exportedAt: { type: string, format: date-time }
        data: { $ref: "#/components/schemas/FamilyOverview" }
        auditEvents:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"